requirements.txt
discord.py
discord.py[voice]

accelerate
bitsandbytes
openai-whisper
OpenCC
sentencepiece
transformers==4.30.2

beautifulsoup4
dlib==19.24.2
googletrans==3.1.0a0
matplotlib
moviepy
numexpr
opencv-python
python-barcode
pydub
qrcode
requests
youtube_dl
